test: 💍 moves test to its own folder
